chore: remove cronjob cmd
