--- pkg/cronjob/audit/interface.go
@@ -1,5 +1,0 @@
-package audit
-
-type ICronjob interface {
-	SyncAuditCycle()
-}
