--- pkg/cronjob/cronjob.go
@@ -1,5 +1,0 @@
-package cronjob
-
-type Cronjob interface {
-	Run()
-}
