+++ pkg/handler/birthday/interface.go
@@ -2,6 +2,6 @@ package birthday
 
 import "github.com/gin-gonic/gin"
 
-type IBirthday interface {
+type IHandler interface {
 	BirthdayDailyMessage(c *gin.Context)
 }
